Move Clear All button to AppBar
