+++ app/src/main/java/com/odysee/app/MainActivity.java
@@ -857,6 +857,37 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
             }
         });
 
+        Context ctx = this;
+        findViewById(R.id.clear_all_library_button).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                AlertDialog.Builder builder = new AlertDialog.Builder(ctx).
+                        setTitle(R.string.confirm_clear_view_history_title).
+                        setMessage(R.string.confirm_clear_view_history)
+                        .setPositiveButton(R.string.yes, new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialogInterface, int i) {
+                                Thread t = new Thread(new Runnable() {
+                                    @Override
+                                    public void run() {
+                                        SQLiteDatabase db = DatabaseHelper.getInstance().getWritableDatabase();
+                                        DatabaseHelper.clearViewHistory(db);
+
+                                        runOnUiThread(new Runnable() {
+                                            @Override
+                                            public void run() {
+                                                ((LibraryFragment) libraryFragment).onViewHistoryCleared();
+                                            }
+                                        });
+                                    }
+                                });
+                                t.start();
+                            }
+                        }).setNegativeButton(R.string.no, null);
+                builder.show();
+            }
+        });
+
         findViewById(R.id.profile_button).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -3301,6 +3332,13 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
         }
     }
 
+    public void switchClearViewHistoryButton(boolean makeItVisible) {
+        if (makeItVisible) {
+            findViewById(R.id.clear_all_library_button).setVisibility(View.VISIBLE);
+        } else {
+            findViewById(R.id.clear_all_library_button).setVisibility(View.GONE);
+        }
+    }
     private boolean isSearchUIActive() {
         return getSupportFragmentManager().findFragmentByTag("SEARCH") != null;
     }
