+++ app/src/main/res/layout/app_bar_main.xml
@@ -182,6 +182,24 @@
                     </LinearLayout>
 
                     <RelativeLayout
+                        android:id="@+id/clear_all_library_button"
+                        android:clickable="true"
+                        android:focusable="true"
+                        android:layout_width="42dp"
+                        android:layout_height="match_parent"
+                        android:background="?attr/selectableItemBackground"
+                        android:visibility="gone">
+
+                        <ImageView
+                            android:layout_width="24dp"
+                            android:layout_height="24dp"
+                            android:layout_centerInParent="true"
+                            android:src="@drawable/ic_clear_all"
+                            android:visibility="visible"
+                            app:tint="@color/actionBarForeground" />
+                    </RelativeLayout>
+
+                    <RelativeLayout
                         android:id="@+id/profile_button"
                         android:clickable="true"
                         android:focusable="true"
